+++ routor/cli.py
@@ -7,8 +7,8 @@ import click
 
 from . import models
 from .engine import Engine
+from .utils import click as click_utils
 from .utils import graph as graph_utils
-from .utils.click import LocationParamType
 
 
 def set_log_level(log_level: Optional[str]) -> None:
@@ -37,7 +37,7 @@ def main() -> None:
     '-e', '--edge-tags', multiple=True, type=str, help="Additional edge tags to fetch."
 )
 @click.argument('location', type=str)
-@click.argument('target', type=click.Path(exists=False, dir_okay=False))
+@click.argument('target', type=click_utils.Path(exists=False, dir_okay=False))
 def download(
     location: str,
     target: Path,
@@ -59,9 +59,9 @@ def download(
 
 @main.command()
 @click.option('--log-level', type=click.Choice(["INFO", "DEBUG"]), default="INFO")
-@click.argument('map_path', type=click.Path(exists=True, dir_okay=False))
-@click.argument('origin', type=LocationParamType())
-@click.argument('destination', type=LocationParamType())
+@click.argument('map_path', type=click_utils.Path(exists=True, dir_okay=False))
+@click.argument('origin', type=click_utils.LocationParamType())
+@click.argument('destination', type=click_utils.LocationParamType())
 @click.argument('weight', type=str)
 def route(
     map_path: Path,