+++ routor/utils/click.py
@@ -1,3 +1,5 @@
+import pathlib
+
 import click
 
 from .. import models
@@ -22,3 +24,12 @@ class LocationParamType(click.ParamType):
             self.fail("Invalid number", param, ctx)
 
         return models.Location(latitude=latitude, longitude=longitude)
+
+
+class Path(click.Path):
+    """
+    A click argument that returns a `pathlib.Path` instance.
+    """
+
+    def convert(self, value, param, ctx) -> pathlib.Path:  # type: ignore
+        return pathlib.Path(super().convert(value, param, ctx))